+++ controller/filestore/file_writer_test.go
@@ -115,7 +115,7 @@ func checkBasicGameJSON(t *testing.T, j string) {
 	require.Equal(t, int32(15), info.Height)
 }
 
-func checkBasicFrameJSON(t *testing.T, j string, turn int64) {
+func checkBasicFrameJSON(t *testing.T, j string, turn int32) {
 	f := &pb.GameFrame{}
 	err := json.Unmarshal([]byte(j), &f)
 	require.NoError(t, err)
@@ -140,7 +140,7 @@ func checkDeadSnakeFrameJSON(t *testing.T, j string) {
 	require.Equal(t, "snake1", f.Snakes[0].ID)
 	require.NotNil(t, f.Snakes[0].Death)
 	require.Equal(t, "death-cause", f.Snakes[0].Death.Cause)
-	require.Equal(t, int64(1), f.Snakes[0].Death.Turn)
+	require.Equal(t, int32(1), f.Snakes[0].Death.Turn)
 }
 
 func TestWriteGameInfo(t *testing.T) {
